refactor(security): relocate findOrCreateColumn, wire index.ts to safe-writes
