+++ __tests__/safe-writes.test.ts
@@ -26,6 +26,7 @@ import {
   writeSafeRichText,
   writeSafeRichTextGrid,
   writeColumn,
+  findOrCreateColumn,
 } from "../src/server/safe-writes";
 
 describe("sanitizeForCell", () => {
@@ -260,5 +261,85 @@ describe("writeColumn", () => {
       ["[SSI Error: AI response contained an external request formula — output rejected]"],
       ["safe"],
     ]);
+  });
+});
+
+describe("findOrCreateColumn", () => {
+  function makeSheet(headers: string[]): GoogleAppsScript.Spreadsheet.Sheet {
+    const values = [headers.slice()];
+    return {
+      getLastColumn: () => headers.length,
+      getRange: jest
+        .fn()
+        .mockImplementation((_row: number, _col: number, numRows?: number, numCols?: number) => {
+          if (numRows === 1 && numCols !== undefined) {
+            return { getValues: () => values };
+          }
+          return { setValue: jest.fn() };
+        }),
+    } as unknown as GoogleAppsScript.Spreadsheet.Sheet;
+  }
+
+  it("returns 1-based index of existing column", () => {
+    const sheet = makeSheet(["Drive Link", "System Prompt", "Output"]);
+    expect(findOrCreateColumn(sheet, "System Prompt")).toBe(2);
+  });
+
+  it("appends new column and returns its 1-based index when not found", () => {
+    const sheet = makeSheet(["Drive Link"]);
+    const setValueMock = jest.fn();
+    (sheet.getRange as jest.Mock).mockImplementation(
+      (_row: number, _col: number, numRows?: number, numCols?: number) => {
+        if (numRows === 1 && numCols !== undefined) {
+          return { getValues: () => [["Drive Link"]] };
+        }
+        return { setValue: setValueMock };
+      },
+    );
+    const idx = findOrCreateColumn(sheet, "New Col");
+    expect(idx).toBe(2);
+    expect(setValueMock).toHaveBeenCalledWith("New Col");
+  });
+
+  it("appends to column 1 when sheet is empty", () => {
+    const setValueMock = jest.fn();
+    const sheet = {
+      getLastColumn: () => 0,
+      getRange: jest.fn().mockReturnValue({ setValue: setValueMock }),
+    } as unknown as GoogleAppsScript.Spreadsheet.Sheet;
+    const idx = findOrCreateColumn(sheet, "My Col");
+    expect(idx).toBe(1);
+    expect(setValueMock).toHaveBeenCalledWith("My Col");
+  });
+
+  it("applies wrapStrategy to the new column range when provided", () => {
+    const setValueMock = jest.fn();
+    const setWrapStrategyMock = jest.fn();
+    const sheet = {
+      getLastColumn: () => 0,
+      getMaxRows: () => 100,
+      getRange: jest
+        .fn()
+        .mockImplementation((_row: number, _col: number, numRows?: number) =>
+          numRows !== undefined
+            ? { setWrapStrategy: setWrapStrategyMock }
+            : { setValue: setValueMock },
+        ),
+    } as unknown as GoogleAppsScript.Spreadsheet.Sheet;
+    const wrapStrategy = "CLIP" as unknown as GoogleAppsScript.Spreadsheet.WrapStrategy;
+    findOrCreateColumn(sheet, "My Col", wrapStrategy);
+    expect(setWrapStrategyMock).toHaveBeenCalledWith(wrapStrategy);
+  });
+
+  it("sanitizes a dangerous new column title before writing", () => {
+    const setValueMock = jest.fn();
+    const sheet = {
+      getLastColumn: () => 0,
+      getRange: jest.fn().mockReturnValue({ setValue: setValueMock }),
+    } as unknown as GoogleAppsScript.Spreadsheet.Sheet;
+    findOrCreateColumn(sheet, '=IMAGE("evil.com")');
+    expect(setValueMock).toHaveBeenCalledWith(
+      "[SSI Error: AI response contained an external request formula — output rejected]",
+    );
   });
 });
